+++ src/app/shared/consts.ts
@@ -1,3 +1,5 @@
+import { BudgetItem } from "./types";
+
 export const MONTHS = [
   "January",
   "February",
@@ -11,4 +13,126 @@ export const MONTHS = [
   "October",
   "November",
   "December",
+];
+
+export const templateForTwo: BudgetItem[] = [
+  {
+    id: 1,
+    itemName: "Person 1 Salary",
+    category: "Salary",
+    type: "income",
+    planned: 28000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 2,
+    itemName: "Person 2 Salary",
+    category: "Salary",
+    type: "income",
+    planned: 25000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 3,
+    itemName: "Rent/Mortgage",
+    category: "Housing",
+    type: "expense",
+    planned: 12000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 4,
+    itemName: "Car Loan",
+    category: "Loan",
+    type: "expense",
+    planned: 3500,
+    totalAmount: 210000,
+    amountPaidOff: 50000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 5,
+    itemName: "Groceries",
+    category: "Groceries",
+    type: "expense",
+    planned: 4500,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 6,
+    itemName: "Utilities (Water, Electric)",
+    category: "Housing",
+    type: "expense",
+    planned: 1200,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 7,
+    itemName: "Car Insurance",
+    category: "Transportation",
+    type: "expense",
+    planned: 3000,
+    paymentMonths: [6, 12],
+    corrections: [],
+  },
+  {
+    id: 8,
+    itemName: "Gas/Fuel",
+    category: "Transportation",
+    type: "expense",
+    planned: 1000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 9,
+    itemName: "Netflix",
+    category: "Entertainment",
+    type: "expense",
+    planned: 129,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 10,
+    itemName: "Spotify Duo",
+    category: "Entertainment",
+    type: "expense",
+    planned: 149,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 11,
+    itemName: "Mobile Subscription (P1)",
+    category: "Subscriptions",
+    type: "expense",
+    planned: 150,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 12,
+    itemName: "Mobile Subscription (P2)",
+    category: "Subscriptions",
+    type: "expense",
+    planned: 150,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
+  {
+    id: 13,
+    itemName: "Savings",
+    category: "Savings",
+    type: "expense",
+    planned: 5000,
+    paymentMonths: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+    corrections: [],
+  },
 ];
